refresh token added, resource-based authorization added, some updates

+++ SocialNetwork/src/services/IdentityService/IdentityService.PL/Controllers/UsersController.cs
@@ -1,5 +1,7 @@
 ﻿using IdentityService.BLL.DTOs.UserDTOs;
 using IdentityService.BLL.Interfaces;
+using IdentityService.PL.Authorization;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace IdentityService.PL.Controllers
@@ -9,10 +11,13 @@ namespace IdentityService.PL.Controllers
     public class UsersController : ControllerBase
     {
         private readonly IUserService userService;
+        private readonly IAuthorizationService authorizationService;
 
-        public UsersController(IUserService userService)
+        public UsersController(IUserService userService,
+                               IAuthorizationService authorizationService)
         {
             this.userService = userService;
+            this.authorizationService = authorizationService;
         }
 
         [HttpGet]
@@ -35,6 +40,13 @@ namespace IdentityService.PL.Controllers
         [HttpPut]
         public async Task<IActionResult> UpdateUserAsync(UpdateUserDTO updateUserDTO)
         {
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, updateUserDTO.Id, Operations.Update);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             var user = await userService.UpdateUserAsync(updateUserDTO);
 
             return Ok(user);
@@ -43,6 +55,13 @@ namespace IdentityService.PL.Controllers
         [HttpDelete]
         public async Task<IActionResult> RemoveUserByIdAsync(string id)
         {
+            var authorizationResult = await authorizationService.AuthorizeAsync(User, id, Operations.Delete);
+
+            if (!authorizationResult.Succeeded)
+            {
+                return Forbid();
+            }
+
             await userService.RemoveUserByIdAsync(id);
 
             return Ok();
